Cleanup: Fix more clang warnings.

+++ src/modules_mod.cpp
@@ -118,7 +118,7 @@ bool flanger_audio_module::get_layers(int index, int generation, unsigned int &l
     layers = LG_REALTIME_GRAPH | (generation ? 0 : LG_CACHE_GRID);
     return true;
 }
-float flanger_audio_module::freq_gain(int subindex, float freq) const
+float flanger_audio_module::freq_gain(int subindex, double freq) const
 {
     return (subindex ? right : left).freq_gain(freq, srate);                
 }
@@ -220,7 +220,7 @@ bool phaser_audio_module::get_layers(int index, int generation, unsigned int &la
     return true;
 }
 
-float phaser_audio_module::freq_gain(int subindex, float freq) const
+float phaser_audio_module::freq_gain(int subindex, double freq) const
 {
     return (subindex ? right : left).freq_gain(freq, srate);                
 }
@@ -627,7 +627,7 @@ bool multichorus_audio_module::get_gridline(int index, int subindex, int phase, 
     return false;
 }
 
-float multichorus_audio_module::freq_gain(int subindex, float freq) const
+float multichorus_audio_module::freq_gain(int subindex, double freq) const
 {
     if (subindex == 2)
         return *params[par_amount] * left.post.freq_gain(freq, srate);
